pyproject.toml
[tool.poetry]
name = "automation"
version = "0.1.0"
description = ""
authors = ["haifeng <1353120323@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lxml = "^5.1.0"
requests = "^2.31.0"
numpy = "^1.26.4"
jsonpath = "^0.82.2"
pandas = "^2.2.0"
tabulate = "^0.9.0"
attrs = "^23.2.0"
pyyaml = "^6.0.1"
ipykernel = "^6.29.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
